refactor: change `[filterId]/edit` to `edit/[filterId]`
